Use prepend a patch module instead of override UserAnswersSerializer
