+++ config/initializers/decidim_override.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../../lib/decidim/forms/user_answers_serializer"
-
 Rails.application.config.to_prepare do
   # Decidim::Proposals::ProposalWizardCreateStepForm
   #
